+++ packages/expander/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @mikqi/expander
+# @kisanak/expander
 
 ## 1.0.0
 
 ### Major Changes
 
-- [`69f4ce5`](https://github.com/mikqi/product-card-monorepo/commit/69f4ce57316a45b647bb149117e5d6d2819fbd96) Thanks [@mikqi](https://github.com/mikqi)! - Initial release
+- [`69f4ce5`](https://github.com/kisanak/product-card-monorepo/commit/69f4ce57316a45b647bb149117e5d6d2819fbd96) Thanks [@kisanak](https://github.com/kisanak)! - Initial release